refactor: unify `ConnectionOptions`, `PoolOptions` and `ConnectionConfig` types
