+++ typings/mysql/index.d.ts
@@ -71,4 +71,19 @@ export function raw(sql: string): {
   toSqlString: () => string;
 };
 
+export interface ConnectionConfig extends ConnectionOptions {
+  mergeFlags(defaultFlags: string[], userFlags: string[] | string): number;
+  getDefaultFlags(options?: ConnectionOptions): string[];
+  getCharsetNumber(charset: string): number;
+  getSSLProfile(name: string): { ca: string[] };
+  parseUrl(url: string): {
+    host: string;
+    port: number;
+    database: string;
+    user: string;
+    password: string;
+    [key: string]: any;
+  };
+}
+
 export function createServer(handler: (conn: BaseConnection) => any): Server;
